% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{compress}
\alias{compress}
\title{Utilities for moving between weighted and unweighted versions of data matrices and fits}
\usage{
compress(data)
}
\arguments{
\item{data}{a numeric matrix}
}
\value{
an object of class compressed.
}
\description{
\link{compress} finds the unique rows of a matrix and returns those along with the respective weight matrix.
}
\details{
asdjasdjkfhlksdjhflas
}
