% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress.R
\name{decompress}
\alias{decompress}
\title{Utilities for moving between weighted and unweighted versions of data matrices and fits}
\usage{
decompress(obj, ...)
}
\arguments{
\item{obj}{An object of class \link{compressed} or of class \link{brRasch}}
}
\description{
\link{compress} finds the unique rows of a matrix and returns those along with the respective weight matrix.
}
