% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{irf.brRasch}
\alias{irf.brRasch}
\title{Plot the item response functions for objects of class brRasch}
\usage{
\method{irf}{brRasch}(obj, item = 0, col = "black", ncol = 1, ...)
}
\arguments{
\item{obj}{a brRasch object}

\item{item}{the item for which the item response function is request. Default value is 0 which plots the item response function for all items}

\item{col}{the colour to be used for plotting (defaults to "black")}

\item{ncol}{the number of columns to be used for the resultant matrix of item response function plots}
}
\description{
adfssfafasdf
}
