% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.brRasch}
\alias{plot.brRasch}
\title{Plot the abilities when \code{dim} is 0 (1PL), \code{dim} is
    1 (2PL) and \code{dim} is 2.}
\usage{
\method{plot}{brRasch}(x, lab = "ability", col = "black", ...)
}
\description{
adfssfafasdf
}

